requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24

passlib~=1.7.4
python-jose~=3.5.0
boto3~=1.40.16
botocore~=1.40.16
openpyxl~=3.1.5
aiohttp~=3.12.15
pytest~=8.4.1
httpx~=0.28.1
alembic~=1.16.4
celery~=5.5.3
starlette~=0.47.2
redis~=6.4.0
psutil~=7.0.0
prometheus_client~=0.22.1
kombu~=5.5.4